impl/refactor documet
