+++ rust-runtime/aws-smithy-types/src/lib.rs
@@ -14,8 +14,7 @@
 )]
 
 use crate::error::{TryFromNumberError, TryFromNumberErrorKind};
-use std::collections::HashMap;
-
+mod document;
 pub mod base64;
 pub mod date_time;
 pub mod endpoint;
@@ -24,6 +23,7 @@ pub mod primitive;
 pub mod retry;
 pub mod timeout;
 
+pub use crate::document::Document;
 pub use crate::date_time::DateTime;
 pub use error::Error;
 
@@ -52,72 +52,6 @@ impl Blob {
 impl AsRef<[u8]> for Blob {
     fn as_ref(&self) -> &[u8] {
         &self.inner
-    }
-}
-
-/* ANCHOR: document */
-
-/// Document Type
-///
-/// Document types represents protocol-agnostic open content that is accessed like JSON data.
-/// Open content is useful for modeling unstructured data that has no schema, data that can't be
-/// modeled using rigid types, or data that has a schema that evolves outside of the purview of a model.
-/// The serialization format of a document is an implementation detail of a protocol.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Document {
-    /// JSON object
-    Object(HashMap<String, Document>),
-    /// JSON array
-    Array(Vec<Document>),
-    /// JSON number
-    Number(Number),
-    /// JSON string
-    String(String),
-    /// JSON boolean
-    Bool(bool),
-    /// JSON null
-    Null,
-}
-
-impl From<bool> for Document {
-    fn from(value: bool) -> Self {
-        Document::Bool(value)
-    }
-}
-
-impl From<String> for Document {
-    fn from(value: String) -> Self {
-        Document::String(value)
-    }
-}
-
-impl From<Vec<Document>> for Document {
-    fn from(values: Vec<Document>) -> Self {
-        Document::Array(values)
-    }
-}
-
-impl From<HashMap<String, Document>> for Document {
-    fn from(values: HashMap<String, Document>) -> Self {
-        Document::Object(values)
-    }
-}
-
-impl From<u64> for Document {
-    fn from(value: u64) -> Self {
-        Document::Number(Number::PosInt(value))
-    }
-}
-
-impl From<i64> for Document {
-    fn from(value: i64) -> Self {
-        Document::Number(Number::NegInt(value))
-    }
-}
-
-impl From<i32> for Document {
-    fn from(value: i32) -> Self {
-        Document::Number(Number::NegInt(value as i64))
     }
 }
